+++ packages/relay-bitcoin-wallet-adapter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @reservoir0x/relay-bitcoin-wallet-adapter
 
+## 16.0.1
+
+### Patch Changes
+
+- Updated dependencies [f851b3c]
+  - @relayprotocol/relay-sdk@5.1.1
+
 ## 16.0.0
 
 ### Patch Changes